% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exp_val.R
\name{exp_val}
\alias{exp_val}
\title{Functions to obtain the expected effect sizes given publication bias
Parameter:
\itemize{
\item PB = Publication Bias... proportion of non-significant effect sizes published
\item g = true effect size as standardized mean difference in form of Hedges' g
\item N = total sample size (N = nT + nC)
\item Zcv = critical Z value (used to calculate the critical value ycv)
\item vg = sampling variance Hedges' g
\item tau2 = (residual) between-study variance (\eqn{\tau^2})
\item I2 = porportion of within- to (residual) between-study variance
\item E_sig = Expected effect size upper truncated normal distribution (significant effect sizes)
\item E_nsig = Expected effect size lower truncated normal distribution (nonsignificant effect sizes)
\item E = expected effect size given publication bias
\item p_ns = probability of a non-significant effect size
Function to obtain the expected value given publication bias for one true effect size g (Hedges' g)
exp_val()
}}
\usage{
exp_val(PB, Zcv, g, N, I2 = NA, tau2 = NA, lower.tail = FALSE)
}
\arguments{
\item{PB}{amount of publication bias between 0 (extreme publication bias) and 1 (no publication bias)}

\item{Zcv}{critical Z value defined by the normal quantile function. For instance Zcv = qnorm(.975).}

\item{g}{True effect size as standardized mean difference (we used Hedges' g)}

\item{N}{Primary study total sample size}

\item{I2}{heterogeneity estimate in the meta-analysis (not necessary when tau2 is specified)}

\item{tau2}{heterogeneity estimate in the meta-analysis (\eqn{\tau^2}) (not necessary when I2 is specified)}

\item{lower.tail}{Indicates sidedness of the effect size testing (e.g., lower.tail = FALSE indicates that one tests for a positive effect.)}
}
\value{
This function returns a single numeric value of the expected effect size given publication bias for a given effect size
}
\description{
Function to obtain the expected value given publication bias for one true effect size g (Hedges' g)
}
