% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exp_val.R
\name{exp_val_continuous}
\alias{exp_val_continuous}
\title{Function to obtain the expected value given publication bias for a true effect size g (Hedges' g)
with a given primary study sample size in a meta-analysis. Compared to the exp_val() function, the
true effect sizes and sample sizes of other studies in the meta-analysis are taken into account when
calculating the residual $$\tau^2$$ for a given residual $$I^2$$.
exp_val_MA}
\usage{
exp_val_continuous(PB, Zcv, N, Nvec, I2, x1, x1vec, beta0, beta1)
}
\arguments{
\item{PB}{amount of publication bias between 0 (extreme publication bias) and 1 (no publication bias)}

\item{Zcv}{critical Z value defined by the normal quantile function. For instance Zcv = qnorm(.975).}

\item{N}{Primary study total sample size}

\item{Nvec}{Vector of all primary study total sample sizes in the meta-analysis}

\item{I2}{residual heterogeneity estimate in the meta-analysis}

\item{x1}{moderator value of the given study in the meta-analysis}

\item{x1vec}{vector with moderator values of all studies in the meta-analysis}

\item{beta0}{true intercept parameter of the meta-analysis}

\item{beta1}{true slope parameter (moderator effect) of the meta-analysis}

\item{g}{True effect size as standardized mean difference (we used Hedges' g)}
}
\value{
This function returns a single numeric value of the expected effect size given publication bias for a given effect size in a meta-analysis.
}
\description{
Function to obtain the expected value given publication bias for a true effect size g (Hedges' g)
with a given primary study sample size in a meta-analysis. Compared to the exp_val() function, the
true effect sizes and sample sizes of other studies in the meta-analysis are taken into account when
calculating the residual $$\tau^2$$ for a given residual $$I^2$$.
}
