% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exp_val.R
\name{exp_val_MA}
\alias{exp_val_MA}
\title{Function to obtain the expected value given publication bias for a true effect size g (Hedges' g)
with a given primary study sample size in a meta-analysis. Compared to the exp_val() function, the
true effect sizes and sample sizes of other studies in the meta-analysis are taken into account when
calculating the residual $$tau^2$$ for a given residual $$I^2$$.
exp_val_MA}
\usage{
exp_val_MA(
  PB,
  Zcv,
  N = NA,
  Nvec = NA,
  vg = NA,
  vgvec = NA,
  I2,
  x1,
  x1vec,
  beta0,
  beta1,
  NoPB = FALSE,
  lower.tail = FALSE
)
}
\arguments{
\item{PB}{amount of publication bias between 0 (extreme publication bias) and 1 (no publication bias)}

\item{Zcv}{critical Z value defined by the normal quantile function. For instance Zcv = qnorm(.975).}

\item{N}{Primary study total sample size}

\item{Nvec}{Vector of all primary study total sample sizes in the meta-analysis (not necessary when vg and vgvec are specified)}

\item{vg}{Primary study sampling variance (not necessary when N and Nvec are specified)}

\item{vgvec}{Vector of all primary study sampling variances in the meta-analysis (not necessary when N and Nvec are specified)}

\item{I2}{residual heterogeneity estimate in the meta-analysis}

\item{x1}{moderator value of the given study in the meta-analysis}

\item{x1vec}{vector with moderator values of all studies in the meta-analysis}

\item{beta0}{true intercept parameter of the meta-analysis}

\item{beta1}{true slope parameter (moderator effect) of the meta-analysis}

\item{NoPB}{whether the given primary study is assumed to be affected by publication bias (FALSE = default) or
whether the study is unaffected by publication bias (TRUE)}

\item{lower.tail}{Indicates sidedness of the effect size testing (e.g., lower.tail = FALSE indicates that one tests for a positive effect.)}
}
\value{
This function returns a single numeric value of the expected effect size given publication bias for a given effect size in a meta-analysis.
}
\description{
Function to obtain the expected value given publication bias for a true effect size g (Hedges' g)
with a given primary study sample size in a meta-analysis. Compared to the exp_val() function, the
true effect sizes and sample sizes of other studies in the meta-analysis are taken into account when
calculating the residual (\eqn{\tau^2}) for a given residual (\eqn{I^2}).
}
