% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exp_val.R
\name{PB_betas}
\alias{PB_betas}
\title{Publication biased intercept and slope parameters in a meta-anaylsis with one moderator}
\usage{
PB_betas(
  dat,
  beta0,
  beta1,
  PP,
  I2 = NA,
  tau2 = NA,
  mods,
  Zcv,
  lower.tail = FALSE
)
}
\arguments{
\item{dat}{data frame with columns vi, and NoPB}

\item{beta0}{true intercept in the meta-analysis}

\item{beta1}{true slope or moderator effect in the meta-analysis}

\item{PP}{publication probability for non-significant studies amount between 0 (extreme publication bias) and 1 (no publication bias)}

\item{I2}{residual heterogeneity indicator in form of I2. Please specify either I2 or tau2, not both.}

\item{tau2}{residual between study variance. Please specify either I2 or tau2, not both.}

\item{mods}{moderator variable (as would be inputted in the metafor mods statement)}

\item{Zcv}{critical Z value for testing the primary study effect sizes}

\item{lower.tail}{Indicates sidedness of the effect size testing (e.g., lower.tail = FALSE indicates that one tests for a positive effect.)}
}
\value{
This function return a vector with the estimated intercept and slope
parameters give publication bias (i.e, c(b0, b1))
}
\description{
calls publication biased data via exp_val_MA() and calculates publication biased betas using betas_PB()
}
