% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{PBanalysis_plots}
\alias{PBanalysis_plots}
\title{Shiny app final default plot}
\usage{
PBanalysis_plots(
  dat,
  mods,
  mem,
  Zcv,
  beta0 = 0,
  beta1 = 0,
  mod.title = "Moderator value",
  heterogeneity = "tau2res",
  I2res = c(0, 0.25, 0.5, 0.75),
  tau2res = c(0, 0.01, 0.04, 0.11),
  PP = c(0, 0.05, 0.2, 0.5, 1),
  lower.tail = FALSE
)
}
\arguments{
\item{dat}{dataframe to be used for plotting}

\item{mods}{moderator variable}

\item{mem}{metafor mixed-effects meta-analysis model}

\item{Zcv}{critical Z value for testing effect sizes}

\item{beta0}{true intercept in the meta-analysis}

\item{beta1}{true slope/moderator effect in the meta-analysis}

\item{mod.title}{character string with the x-axis label specifying the moderator (default = "Moderator value")}

\item{heterogeneity}{indicator whether the figures should be created for four different levels of tau2res or I2res. Possible values are "tau2res" or "I2res".}

\item{I2res}{vector with four levels of residual I2 (default is I2res=c(0, 0.25, 0.5, 0.75))}

\item{tau2res}{vector with four between study variances (default is tau2res = c(0, 0.01, 0.04, 0.11) based on van Erp et al. (2017))}

\item{PP}{vector with four publication probabilities for non-significant studies (PP=c(0, 0.05, 0.2, 0.5, 1))}

\item{lower.tail}{Indicates sidedness of the effect size testing (e.g., lower.tail = FALSE indicates that one tests for a positive effect.)}
}
\value{
Shiny app final default plot
}
\description{
Function to create the shiny app default plot used in sensitivity publication bias analyses for moderator effects in meta-analysis.
}
