% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{Plot_additional_analysis}
\alias{Plot_additional_analysis}
\title{Shiny app additional plot}
\usage{
Plot_additional_analysis(
  data,
  beta0,
  beta1,
  I2,
  PB,
  mem,
  betasPB,
  mod.title = "Moderator value"
)
}
\description{
Shiny app additional plot
}
