% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_grid_1legend}
\alias{plot_grid_1legend}
\title{Generates a figure with 4 sub figures and one legend}
\usage{
plot_grid_1legend(p1, p2, p3, p4)
}
\arguments{
\item{p1}{first ggplot figure}

\item{p2}{second ggplot figure}

\item{p3}{third ggplot figure}

\item{p4}{fourth ggplot figure}
}
\value{
a ggplot figure
}
\description{
Generates a figure with 4 sub figures and one legend
}
