% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_grid_1legend_2p}
\alias{plot_grid_1legend_2p}
\title{Generates Figure with 2 sub Figures in a row and one legend}
\usage{
plot_grid_1legend_2p(p1, p2, legend)
}
\description{
Generates Figure with 2 sub Figures in a row and one legend
}
