% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{PublicationBiasInModeratorAnalysis}
\alias{PublicationBiasInModeratorAnalysis-package}
\alias{PublicationBiasInModeratorAnalysis}
\title{Package-level imports}
\description{
Provides functions and a Shiny app for analyzing publication bias in moderator analyses. Includes methods for expected value calculations, plotting, and simulation.
}
\author{
\strong{Maintainer}: Franziska Rüffer \email{f.f.ruffer@gmail.com}

}
\keyword{internal}
