% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{p_bias_diff_exp_val}
\alias{p_bias_diff_exp_val}
\title{Plotting Function Figure 2}
\usage{
p_bias_diff_exp_val(data, title, size_text = 8, axis_text = 23)
}
\description{
Function to create the subplot of Figure 2
}
