% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{p_bias_exp_val}
\alias{p_bias_exp_val}
\title{Plotting Function Figure 1}
\usage{
p_bias_exp_val(data, x, y, title, size_text = 8, axis_text = 23)
}
\description{
Function to create the subplots of Figure 1
}
