% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exp_val.R
\name{PB_betas}
\alias{PB_betas}
\title{Publication biased intercept and slope parameters in a meta-anaylsis with one moderator}
\usage{
PB_betas(dat, beta0, beta1, PB, I2, mods, Zcv)
}
\arguments{
\item{dat}{data frame with columns vi, and NoPB}

\item{beta0}{true intercept in the meta-analysis}

\item{beta1}{true slope or moderator effect in the meta-analysis}

\item{PB}{publication bias (0 = extreme publication bias, 1 = no publication bias)}

\item{I2}{residual heterogeneity indicator in form of I2}

\item{mods}{moderator variable (as would be inputted in the metafor mods statement)}

\item{Zcv}{critical Z value for testing the primary study effect sizes}
}
\value{
This function return a vector with the estimated intercept and slope
parameters give publication bias (i.e, c(b0, b1))
}
\description{
calls publication biased data via exp_val_MA() and calculates publication biased betas using betas_PB()
}
