% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{PBanalysis_plots}
\alias{PBanalysis_plots}
\title{Shiny app final default plot}
\usage{
PBanalysis_plots(
  dat,
  mods,
  mem,
  Zcv,
  beta0 = 0,
  beta1 = 0,
  mod.title = "Moderator value",
  I2res = c(0, 0.25, 0.5, 0.75),
  PB = c(0, 0.05, 0.2, 0.5, 1)
)
}
\description{
Shiny app final default plot
}
