% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_grid_1legend}
\alias{plot_grid_1legend}
\title{Generates Figure with 4 sub Figures and one legend}
\usage{
plot_grid_1legend(p1, p2, p3, p4, legend)
}
\description{
Generates Figure with 4 sub Figures and one legend
}
