% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_grid_1legend_2p}
\alias{plot_grid_1legend_2p}
\title{Generates a figure with two sub figures in a row and one legend
#' @description Generates a figure with 2 sub figures and one legend}
\usage{
plot_grid_1legend_2p(p1, p2)
}
\arguments{
\item{p1}{first ggplot figure}

\item{p2}{second ggplot figure}
}
\value{
a ggplot figure
}
\description{
Generates a figure with two sub figures in a row and one legend
#' @description Generates a figure with 2 sub figures and one legend
}
