% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{_PACKAGE}
\name{PublicationBiasInModeratorAnalysis}
\alias{PublicationBiasInModeratorAnalysis}
\title{Package-level imports}
\description{
Publication bias sensitvity analyses for moderator effects in meta-analyses. R functions under MIT license; example figures and paper under CC-BY-4.0.
}
\author{
\strong{Maintainer}: Franziska Rüffer \email{f.f.ruffer@gmail.com}

}
\keyword{internal}
