% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exp_val.R
\name{sol.g2}
\alias{sol.g2}
\title{Solves for the true effect size g2 in group2, s.t. E1=E2 (hidden moderator effect)}
\usage{
sol.g2(N1, N2, g1, Zcv, PB1, PB2, I2 = 0)
}
\description{
Solves for the true effect size g2 in group2, s.t. E1=E2 (hidden moderator effect)
}
